fixes [SetBpm] by using the BpmProvider Instance directly
BpmProvider and TapProvider moved to core
interfaces are no longer needed, but i haven't discarded them yet
