+++ Operators/Lib/numbers/anim/vj/SetBpm.cs
@@ -1,6 +1,5 @@
-using Lib.Utils;
 using T3.Core.Animation;
-using T3.Core.Audio;
+using T3.Core.IO;
 using T3.Core.Utils;
 
 namespace Lib.numbers.anim.vj;
